create function test

+++ src/include/optimizer/logical_operators.h
@@ -1031,7 +1031,7 @@ class LogicalCreateFunction : public OperatorNode<LogicalCreateFunction> {
                        std::string function_name, parser::PLType language, std::vector<std::string> &&function_body,
                        std::vector<std::string> &&function_param_names,
                        std::vector<parser::BaseFunctionParameter::DataType> &&function_param_types,
-                       parser::BaseFunctionParameter::DataType return_type, int param_count, bool replace);
+                       parser::BaseFunctionParameter::DataType return_type, size_t param_count, bool replace);
 
   bool operator==(const BaseOperatorNode &r) override;
   common::hash_t Hash() const override;
@@ -1086,7 +1086,7 @@ class LogicalCreateFunction : public OperatorNode<LogicalCreateFunction> {
   /**
    * @return number of parameters of the user defined function
    */
-  int GetParamCount() const { return param_count_; }
+  size_t GetParamCount() const { return param_count_; }
 
  private:
   /**
@@ -1137,7 +1137,7 @@ class LogicalCreateFunction : public OperatorNode<LogicalCreateFunction> {
   /**
    * Number of parameters
    */
-  int param_count_ = 0;
+  size_t param_count_ = 0;
 };
 
 /**
